Passage des coordonnées en double

+++ moovsimply/src/main/java/fr/duchesses/moov/models/Coordinates.java
@@ -1,19 +1,19 @@
 package fr.duchesses.moov.models;
 
 public class Coordinates {
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
-    public Coordinates(long latitude, long longitude) {
+    public Coordinates(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 }
